remove anthropic, batch, move gpt to gpt.py
